feat:Issue #2 Initial DB setup
Issue #2 Got the initial app building and running

Have a Postgres DB in Docker, initialized with some dummy data, and
confirmed I can connect to the database and read it. Next steps are
going to be looking at moving DB-specific stuff to a package for proper
code organization, and starting on implementing user management.