Issue #4 Got traces and spans working
Can't do it as a middleware, which makes sense, but is a little sad. I
was always going to have to do some instrumentation at the handler level
anyways, so I guess it's no big deal. Putting the arbitrarily wide
attributes on the span instead of a canonical log line (REALLY wanted to
experiment with that but this seems to be the preferred way to do things
in OTel)