Issue #11 Fixed the parallelization problem with the tests
Removed the .Close() test since I'm using the method to resolve the test
issues. If it's not working the other tests will fail. Debugging the
migrations logic now. I got distracted with the tests so I had forgotten
when I was on implementing the logic for reading the files from the disk
in a testable way, but I have a TODO in there now for a reminder