Issue #11 Tests run through
I thought creating a table would leave a rows impacted count (if only on
the list of tables, but apparently not. Also need to write a query for
the migrations table to confirm I see my file name(s) there when
migrations succeed. I still wish I could use embed, but I don't know a
way to inject a test without creating a new helper function that takes
an explicit FS parameter, just for testing, which seems a waste.