Issue #15 Got messed, so starting over.
Wrote a quick recollection of what the index handler test was. Also
moved creating the observability container to just the health check test
(which is the only place it's used), and left the DB container in it's
helper function. May move the IndexHandler and HealthCheckHandler to
application_handers_test and just leave Shutdown() here (and probably
add Run())