Issue #15 Basic signup form
Got a simple form in the UI, need to install HTMX so I can call the
back-end to create the account. Also want to re-think my person schema
to make sure I have the right fields in my form. Also moved the data
directories for Grafana and Postgres out of the docker/ directory
because that was causing problems. Need to write an init script to
populate these things so this is easy to set up.