Issue #15 Moving forward, but major refactoring needed soon.
Immediate term issue - showing the signup form on the index page means
tweaking logic in the template for 1 place breaks it in the other -
going to need to either trigger a redirect or show links to take you to
the correct page.
Medium term (next issue) - Either refactor databasing or go LibSQL. Had
to bump test DB container timeout because 5 stopped cutting it on tests.
Either full refactor so there's a database per package test (with
different schemas for individual tests), or just use a SQLite file
locally for everything (more consistent, and just seems like a better
fit the more I think about it)