feat: #20 Database updates for households (#33)
Added a household and household_person table for holding the household name and managing a many-to-one relationship between households and the people in them. Updated the profile page to include the household name (members of the household can change the name). Also added the ability to update the household from the profile edit form.