feat: #29 Add managed profiles
For children and other users who for any reason can't (or won't) log in
to manage their registry on their own.

Updated the profile page to be based off a list of profiles (so managed
profiles can be added to the end of the list). Updates are still broken,
I need to adjust where the range lives relative to the cards. Also
updated migrations so that a single migration failure doesn't stop other
files from running (NEEDS A TEST UPDATE TO VALIDATE). Also allowed
multiple statments in 1 migration file. The file will still fail if ANY
statement is invalid (NEED A NEW TEST TO VALIDATE THAT TOO).

fix: #29 Fixed the broken tests

Moved the loop to the profile_page.html, and the profile_form.html
template now just holds a single card. Also fixed a bug in the migration
logic.

fix: #29 Fixed the quotes in the type column schema change

feat: #29 Updated tests to create managed profiles in the main profile's
household

This is a prepatory step for testing managed people in the profile page.
Also added test case for verifying that loading a profile page pulls in
managed profiles for that household. Still need to write the test case
validating that updating a managed profile updates just that profile.

feat: #29 Started work on querying managed profiles on profile page load

Still need to debug something (managed profiles are pulling data from
the parent record), but the query for pulling the managed profiles are
there. Need to decide if I want the indicator for a managed profile to
be type = 'MANAGED' or type <> 'NORMAL'

fix: #29 Fixed profile read bug.

Caused by pulling email address (which is NULL in the DB) and trying to
read to a string in the struct. Avoidable problem, but did read up on
addressing that in the future.

feat #29: Almost done with managed profiles

Some last little bugs around ensuring emails are unique when they can
also not have data. I have a StackOverflow answer that will probably
work, but it's late so I'll test it some time tomorrow.

fix: #29 Tests are passing

fix: #29 Initial rendering of managed profiles works

Some bugs when saving a managed profile, started working on those, but
have some query errors that need debugging later. Also need to figure
out if I should update element checks to flag an element being on the
page that shouldn't (maybe co-opt the visible check -- do I care about
on the page and now shown versus not there? I might)

fix: #29 Working on update bugs

Test is failing but lookup query shows the test is looking up the normal
person account, not the managed profile. Need to fix that lookup query
to pull the actually updated profile.

fix: #29 Fixed updating child profiles